+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - 对项目中的单级日志相关功能提供了框架级别的功能更新。
 - BehaviorAnalyse AOP 功能增强。
+- 优化 com.dwarfeng.subgrade.sdk.jdbc.SQLAndParameter 的代码结构。
 
 #### Bug修复